pyproject.toml
[project]
name = "jira-langgraph-agent"
version = "0.1.0"
description = "A LangGraph-powered AI agent for interacting with JIRA tickets using Google Gemini"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-google-genai>=0.1.3",
    "langgraph>=0.0.52",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "jira>=3.5.0",
    "pandas>=2.0.0",
    "pdfplumber>=0.10.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "jinja2>=3.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
]
